+++ packages/components/src/components/accordion/accordion-utils.ts
@@ -1,3 +1,5 @@
+import type { HeadingTag } from '../../utils';
+
 export const ACCORDIONS_BACKGROUNDS = ['canvas', 'frosted', 'none'] as const;
 export type AccordionBackground = (typeof ACCORDIONS_BACKGROUNDS)[number];
 
@@ -5,3 +7,6 @@ export const ACCORDION_ALIGN_ICONS = ['start', 'end'] as const;
 export type AccordionAlignIcon = (typeof ACCORDION_ALIGN_ICONS)[number];
 
 export type AccordionUpdateEventDetail = { open: boolean };
+
+/** @deprecated */
+export type AccordionHeadingTag = HeadingTag;